requirements.txt
Flask==3.0.0
faiss-cpu==1.7.4
chromadb==0.4.22
pinecone-client==3.2.2
ollama==0.1.8
langchain==0.1.15
pypdf==3.16.3
transformers==4.38.2
torch==2.1.0
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0  # For production deployment
